+++ components/frontend/src/widgets/buttons/ActionAndItemPickerButton.jsx
@@ -1,5 +1,5 @@
 import ArrowDropDownIcon from "@mui/icons-material/ArrowDropDown"
-import { Button, Menu, MenuItem, Tooltip, Typography } from "@mui/material"
+import { Button, MenuItem, MenuList, Popover, TextField, Tooltip, Typography } from "@mui/material"
 import { element, func, string } from "prop-types"
 import { useState } from "react"
 
@@ -7,11 +7,14 @@ import { ItemBreadcrumb } from "../ItemBreadcrumb"
 
 export function ActionAndItemPickerButton({ action, itemType, onChange, getOptions, icon }) {
     const [anchorEl, setAnchorEl] = useState()
+    const [query, setQuery] = useState("") // Search query to filter items
     const handleMenu = (event) => setAnchorEl(event.currentTarget)
     const onClick = (value) => {
         onChange(value)
         setAnchorEl(null)
     }
+
+    const options = getOptions().filter((option) => option.text.toLowerCase().includes(query.toLowerCase()))
 
     const breadcrumbProps = { report: "report" }
     if (itemType !== "report") {
@@ -31,30 +34,30 @@ export function ActionAndItemPickerButton({ action, itemType, onChange, getOptio
                     <ArrowDropDownIcon />
                 </Button>
             </Tooltip>
-            <Menu
-                id="action-menu"
-                anchorEl={anchorEl}
-                onClose={() => setAnchorEl(null)}
-                open={Boolean(anchorEl)}
-                slotProps={{
-                    paper: {
-                        style: {
-                            maxHeight: 250,
-                        },
-                    },
-                }}
-            >
-                <MenuItem disabled divider>
-                    <Typography variant="button">
-                        <ItemBreadcrumb {...breadcrumbProps} />
-                    </Typography>
-                </MenuItem>
-                {getOptions().map((option) => (
-                    <MenuItem key={option.key} onClick={() => onClick(option.value)}>
-                        {option.content}
+            <Popover id="dropdown-menu" anchorEl={anchorEl} onClose={() => setAnchorEl(null)} open={Boolean(anchorEl)}>
+                <TextField
+                    autoFocus
+                    fullWidth
+                    label={`Filter ${itemType}s`}
+                    onChange={(event) => setQuery(event.target.value)}
+                    value={query}
+                    variant="outlined"
+                    sx={{ ml: "10px", mt: "10px", pr: "20px" }}
+                    type="search"
+                />
+                <MenuList sx={{ height: "30vh", width: "50vw" }}>
+                    <MenuItem disabled divider>
+                        <Typography variant="button">
+                            <ItemBreadcrumb {...breadcrumbProps} />
+                        </Typography>
                     </MenuItem>
-                ))}
-            </Menu>
+                    {options.map((option) => (
+                        <MenuItem key={option.key} onClick={() => onClick(option.value)}>
+                            {option.content}
+                        </MenuItem>
+                    ))}
+                </MenuList>
+            </Popover>
         </>
     )
 }